Enables GTFS aggregation in Damu (#439)
* Adds GTFS merger code.

* Fixes build errors

* Updates paths to zip files in tests

* Sets up Testcontainer for pubsub

Unit tests are now using a testcontainer for Google PubSub instead of depending on emulator standalone

* Adds testcontainer deps to pom

* Removes old references to Marduk

* Fixes failing tests

* Adds new route for GTFS aggregation

* Adds missing constant to GtfsAggregationQueueRouteBuilderTest

* Adds upload of merged gtfs to aggregation route

* Makes aggregation pubsub route work!

* Ensures status updates are published to marduk's pubsub

* Updates delimiter from , => DELIMITER

Temporary fix of delimiter issue. Ideally, we should use , or \n as a delimiter, however this seems to result in encoding issues between marduk and damu, which need to be resolved first.

* Switches to using comma as delimiter

* Ensures gtfs dir is created if non-existent

Also adds some more logging to aggregation route

* Adds aggregation statuses

Ensures damu sends status messages to marduk in the process of merging and uploading merged GTFS files

* Adds cleanup of local working directories

* Ensures JOB_ACTION header is set in test

* Also adds separate tests for basic/extended GTFS conversion headers such that both paths are tested

* Refactoring + testing of aggregation status

* Moves GTFS aggregation into Processors

* Adds marduk pubsub project id to helm

* Removes unnecessary process from test

* Removes unused constants

* Formats code with prettier

Co-authored-by: eibakke <eivind.morris.bakke@entur.org>

* [WIP] Sequential GTFS aggregation Damu after some tests that OOMed.

Co-authored-by: erlendev <erlend.vige@gmail.com>

* Simplifies GTFS aggregation

GTFS aggregation is being simplified because we no longer need to pass separate messages from marduk to damu when aggregating basic and extended gtfs formats. This makes damu agnostic to the job_action header which has previously been used to handle both types of messages.

This commit also adds missing correlation id and job_action headers to the damu.pubsub.http aggregation test message.

* Fixes formatting issue

* Fixes FILE_NAME header overwrite bug

Co-authored-by: eibakke <eivind.morris.bakke@entur.org>

* Removes usage of JOB_ACTION header

* Removes JOB_ACTION header from http script

* Ensure we send failure status messages over pubsub

* Adds dispatch route for GTFS ops

With this commit, messages for triggering export and aggregation of GTFS will be handled by a single Camel route that dispatches the message accordingly

* Fixes failing tests

* Fixes missing formatting

* Improves naming of constants

* Moves upload of gtfs into direct:aggregateGtfs

* Implements file exclusion in (basic) gtfs aggregation

* Explicitly sets handled(true) in aggregation onException

* Refactors AggregationProcessor classes

* Sets handled(true) for GTFS export exceptions

---------

Co-authored-by: erlendev <erlend.vige@gmail.com>